backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
pandas
psycopg2-binary
openpyxl
python-dotenv
gunicorn
xlrd
